chore: remove unused code, move code into correct directory
